Extend the acquisition-offset grid past -4 and to half steps (#3319)
#3318 shipped ACQUISITION_INCLUSION_OFFSET = -3 with -4 tied with it on every
endpoint and the trend not turned, so -4 was the edge of the grid rather than an
optimum. This adds the arms that can find the turn, and the resolution to see it
if it sits between two integers.

The knob is a log2 scale -- inclusion_cost_weights is 2**-k -- so one step
DOUBLES the evidence demanded, and the integer grid is coarse by construction:
between -3 and -4 the cut jumps from 8:1 to 16:1 evidence with nothing in
between, and #3318's own CIs for the two overlap. Half steps are a factor of
sqrt(2) and are real operating points, verified distinct on the shipped
FoldAnchoredCut path before this grid was written -- they survive
snap_cut_to_sample rather than collapsing onto their integer neighbour, which
would have made them silent duplicates of it.

The arithmetic was always continuous, so accepting them is annotations plus one
env parser, not a behaviour change; production keeps the integer -3.

* thresholds.py / voting_iterations.py: widen the inclusion annotations to
  float, and document what a value MEANS. Because the loss is a weighted sum of
  RATES, each normalised by its own class, prevalence divides out --
  rate_crossing puts the prior-odds factor back into lam precisely so the
  realised cut does not carry it -- leaving a log2 likelihood-ratio threshold:
  include x iff f_pos(x)/f_neg(x) > 2**-k. So k=0 is the neutral-evidence point
  and each step of the knob is one bit of evidence. That is the argument for
  parameterising this as an offset at all: a constant shift in evidence-bits is
  prior-free, which is why it transfers across environments where an absolute
  cut would not.
* experiment_config.py: parse CALIB_ACQ_INCLUSION_OFFSET as a float. Parsed as
  an int, "-3.5" raises at import and kills the arm at cell 0; parsed with a
  silent truncation it would be far worse -- the arm would run, complete, and be
  a duplicate of -3.
* analyze_acq.py: derive ARM_K/ARM_LABEL from the arm NAME (an h suffix is a
  half step) and take the arm list from ACQ_ANALYZE_ARMS, defaulting to #2877's
  seven so that run still reproduces. A hand-maintained dict leaves a new arm
  off the frontier's x-axis with no complaint.
* launch_acq_3319.sh: #2877's environment verbatim -- the value of extending a
  grid is that the new arms are comparable to the old ones -- with twelve arms
  and an ACQ_DEEP=1 wave at the 400-click horizon nothing behind this constant
  has ever been measured at.
* PLAN_3319.md: pre-registered, before any arm cell existed.

Refs #3319, #2876, #2877, #2910, #3318.

+++ vtscore/eval/voting_iterations.py
@@ -3753,7 +3753,7 @@ def simulate_voting_iterations(  # noqa: C901
     cut_inclusion_ks: Optional[list[int]] = None,
     cut_inclusion_sink: Optional[list[dict[str, Any]]] = None,
     cut_inclusion_qtilt_steps: Optional[list[float]] = None,
-    acq_inclusion_offset: int = ACQUISITION_INCLUSION_OFFSET,
+    acq_inclusion_offset: float = ACQUISITION_INCLUSION_OFFSET,
     acq_rank_percentile: Optional[float] = None,
     startup_schedule: Optional[str] = None,
     pick_sink: Optional[list[dict[str, Any]]] = None,
